IMS-VT: Clear video pause request list in tracker on call hold.
In case of answering a waiting video call on top of an ongoing
video call, first the ongoing call is paused and then moved to
held state. Later when the held call is swapped back to active,
there is no rtp resume request sent and hence the videopause
tracker will think that the call is still in paused state.
In this state, if call is moved to background, pause request is
filtered out at VideoPauseTracker. To fix this, clear the pause
requests list in VideoPauseTracker when any call is moved to held
state.

Change-Id: I114040a5c94000693a33c6383d33d5618f9f4f44
CRs-Fixed: 2046005